split off metricData and zipper
this should be better than having all this stuff in seemingly arbitrary places in main.go
