+++ main.go
@@ -3,12 +3,9 @@ package main
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"expvar"
 	"flag"
 	"fmt"
-	"io/ioutil"
-	"math"
 	"net/http"
 	_ "net/http/pprof"
 	"net/url"
@@ -23,26 +20,8 @@ import (
 
 	"github.com/bradfitz/gomemcache/memcache"
 	ecache "github.com/dgryski/go-expirecache"
-	pickle "github.com/kisielk/og-rek"
 	"github.com/peterbourgon/g2g"
 )
-
-type metricData struct {
-	pb.FetchResponse
-
-	// extra options
-	drawAsInfinite bool
-	secondYAxis    bool
-	dashed         bool // TODO (ikruglov) smth like lineType would be better
-	color          string
-}
-
-type zipper struct {
-	z      string
-	client *http.Client
-}
-
-var Zipper zipper
 
 var Metrics = struct {
 	Requests         *expvar.Int
@@ -80,163 +59,12 @@ var defaultTimeZone = time.Local
 
 var logger mlog.Level
 
-func (z zipper) Find(metric string) (pb.GlobResponse, error) {
-
-	u, _ := url.Parse(string(z.z) + "/metrics/find/")
-
-	u.RawQuery = url.Values{
-		"query":  []string{metric},
-		"format": []string{"protobuf"},
-	}.Encode()
-
-	var pbresp pb.GlobResponse
-
-	err := z.get("Find", u, &pbresp)
-
-	return pbresp, err
-}
-
-var errNoMetrics = errors.New("no metrics")
-
-func (z zipper) Render(metric string, from, until int32) (metricData, error) {
-
-	u, _ := url.Parse(string(z.z) + "/render/")
-
-	u.RawQuery = url.Values{
-		"target": []string{metric},
-		"format": []string{"protobuf"},
-		"from":   []string{strconv.Itoa(int(from))},
-		"until":  []string{strconv.Itoa(int(until))},
-	}.Encode()
-
-	var pbresp pb.MultiFetchResponse
-	err := z.get("Render", u, &pbresp)
-	if err != nil {
-		return metricData{}, err
-	}
-
-	if m := pbresp.Metrics; len(m) == 0 {
-		return metricData{}, errNoMetrics
-	}
-
-	mdata := metricData{FetchResponse: *pbresp.Metrics[0]}
-
-	return mdata, nil
-}
-
-func (z zipper) Passthrough(metric string) ([]byte, error) {
-
-	u, _ := url.Parse(string(z.z) + metric)
-
-	resp, err := z.client.Get(u.String())
-	if err != nil {
-		return nil, fmt.Errorf("http.Get: %+v", err)
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("ioutil.ReadAll: %+v", err)
-	}
-
-	return body, nil
-}
-
-type unmarshaler interface {
-	Unmarshal([]byte) error
-}
-
-func (z zipper) get(who string, u *url.URL, msg unmarshaler) error {
-	resp, err := z.client.Get(u.String())
-	if err != nil {
-		return fmt.Errorf("http.Get: %+v", err)
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("ioutil.ReadAll: %+v", err)
-	}
-
-	err = msg.Unmarshal(body)
-	if err != nil {
-		return fmt.Errorf("proto.Unmarshal: %+v", err)
-	}
-
-	return nil
-}
+var Zipper zipper
 
 var Limiter limiter
 
 // for testing
 var timeNow = time.Now
-
-func marshalJSON(results []*metricData) []byte {
-
-	var b []byte
-
-	b = append(b, '[')
-
-	var topComma bool
-	for _, r := range results {
-		if r == nil {
-			continue
-		}
-
-		if topComma {
-			b = append(b, ',')
-		}
-		topComma = true
-
-		b = append(b, `{"target":`...)
-		b = strconv.AppendQuoteToASCII(b, r.GetName())
-		b = append(b, `,"datapoints":[`...)
-
-		var innerComma bool
-		t := r.GetStartTime()
-		for i, v := range r.Values {
-			if innerComma {
-				b = append(b, ',')
-			}
-			innerComma = true
-
-			b = append(b, '[')
-
-			if r.IsAbsent[i] || math.IsInf(v, 0) {
-				b = append(b, "null"...)
-			} else {
-				b = strconv.AppendFloat(b, v, 'f', -1, 64)
-			}
-
-			b = append(b, ',')
-
-			b = strconv.AppendInt(b, int64(t), 10)
-
-			b = append(b, ']')
-
-			t += r.GetStepTime()
-		}
-
-		b = append(b, `]}`...)
-	}
-
-	b = append(b, ']')
-
-	return b
-}
-
-func marshalProtobuf(results []*metricData) []byte {
-	response := pb.MultiFetchResponse{}
-	for _, metric := range results {
-		response.Metrics = append(response.Metrics, &((*metric).FetchResponse))
-	}
-	b, err := response.Marshal()
-	if err != nil {
-		logger.Logf("proto.Marshal: %v", err)
-	}
-
-	return b
-}
 
 func writeResponse(w http.ResponseWriter, b []byte, format string, jsonp string) {
 
@@ -268,96 +96,6 @@ func writeResponse(w http.ResponseWriter, b []byte, format string, jsonp string)
 		w.Header().Set("Content-Type", contentTypePNG)
 		w.Write(b)
 	}
-}
-
-func marshalRaw(results []*metricData) []byte {
-
-	var b []byte
-
-	for _, r := range results {
-
-		b = append(b, r.GetName()...)
-
-		b = append(b, ',')
-		b = strconv.AppendInt(b, int64(r.GetStartTime()), 10)
-		b = append(b, ',')
-		b = strconv.AppendInt(b, int64(r.GetStopTime()), 10)
-		b = append(b, ',')
-		b = strconv.AppendInt(b, int64(r.GetStepTime()), 10)
-		b = append(b, '|')
-
-		var comma bool
-		for i, v := range r.Values {
-			if comma {
-				b = append(b, ',')
-			}
-			comma = true
-			if r.IsAbsent[i] {
-				b = append(b, "None"...)
-			} else {
-				b = strconv.AppendFloat(b, v, 'f', -1, 64)
-			}
-		}
-
-		b = append(b, '\n')
-	}
-	return b
-}
-
-func marshalCSV(results []*metricData) []byte {
-
-	var b []byte
-
-	for _, r := range results {
-
-		step := r.GetStepTime()
-		t := r.GetStartTime()
-		for i, v := range r.Values {
-			if !r.IsAbsent[i] {
-				b = append(b, '"')
-				b = append(b, r.GetName()...)
-				b = append(b, '"')
-				b = append(b, ',')
-				b = append(b, time.Unix(int64(t), 0).Format("2006-01-02 15:04:05")...)
-				b = append(b, ',')
-				b = strconv.AppendFloat(b, v, 'f', -1, 64)
-				b = append(b, '\n')
-			}
-			t += step
-		}
-	}
-	return b
-}
-
-func marshalPickle(results []*metricData) []byte {
-
-	var p []map[string]interface{}
-
-	for _, r := range results {
-		values := make([]interface{}, len(r.Values))
-		for i, v := range r.Values {
-			if r.IsAbsent[i] {
-				values[i] = pickle.None{}
-			} else {
-				values[i] = v
-			}
-
-		}
-		p = append(p, map[string]interface{}{
-			"name":   r.GetName(),
-			"start":  r.GetStartTime(),
-			"end":    r.GetStopTime(),
-			"step":   r.GetStepTime(),
-			"values": values,
-		})
-	}
-
-	var buf bytes.Buffer
-
-	penc := pickle.NewEncoder(&buf)
-	penc.Encode(p)
-
-	return buf.Bytes()
 }
 
 const (
